requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cloudinary==1.44.1
contourpy==1.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==46.0.1
cycler==0.12.1
Django==4.2.24
django-cloudinary-storage==0.3.0
django-colorfield==0.14.0
django-cors-headers==4.8.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
Faker==37.8.0
fonttools==4.59.2
gunicorn==23.0.0
idna==3.10
importlib_resources==6.5.2
itypes==1.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.3
matplotlib==3.9.4
mimesis==12.1.0
numpy==2.0.2
openpyxl==3.1.5
packaging==25.0
pillow==11.3.0
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
zipp==3.23.0
